Extract LeaveAccrualService from LeaveBalanceService
The monthly-accrual engine (walk each month, prorate by attendance, post
ACCRUAL transactions, refresh accrued/used/pending) was tangled into
LeaveBalanceService alongside the balance-query and summary logic. Move the
six-method accrual cluster into a focused LeaveAccrualService.

LeaveBalanceService injects it and calls leaveAccrualService.recalculate(balance)
at the three sites that previously called the private recalculateBalance. The
new methods carry no transaction annotation, so they run inside the caller's
existing transaction exactly as before; the managed-entity and save/flush
semantics are unchanged. LeaveBalanceService drops the now-unused
attendance and leave-request repositories.
